fix(lambda-python): bundling with poetry is broken
It looks like something was changed in the base image and there is no
longer write access to the `/tmp` directory which causes bundling with
poetry to fail (see linked issue). This PR updates the Dockerfile to
create a new cache location for both `pip` and `poetry` and switches to
using a virtualenv for python so that it is no longer using root.

To test this I executed the `integ.function.poetry` integration test
both before (to reproduce the error) and after the fix. I'm actually not
sure why our integration tests didn't start failing in the pipeline. The
only thing I can think of is that we are caching the docker images and
it just hasn't pulled down a newer one that has this issue.

fixes #21867