fix(lambda-python): poetry bundling fails on python3.7
For some reason when using the `python3.7` docker image the `useradd`
command fails `/bin/sh: /sbin/useradd: No such file or directory`.

I'm not sure why it works on 3.8 & 3.9 and not 3.7, but since that is
not necessary I'm just removing it.

I've added python3.7 to the integration test to confirm that it works
for 3.7, 3.8, & 3.9