fix(lambda-python): bundling artifacts are written to the entry path
When performing function bundling sometimes the bundling steps will
write files. Currently all these commands are run from the `entry` which
is the actually source code location. This leads to contaminating the
source code with bundling artifacts.

This PR re-orders the bundling steps to first move the `entry` to the
`outputDir` and then perform the bundling steps there.

I've also updated all of the integration tests to use the new
integration test framework and assertions.

fixes #19231