fix(lambda-event-sources): cannot add sqs event source to an imported function
If an SQS event sources is added to an imported function it will throw
an error if the function is not imported with an IAM role.

This PR updates the logic to only attempt to add permissions to the
principal if the role exists, otherwise it will add a warning indicating
that permissions were not added.

fixes #12607