fix(bootstrap): allow tags to be updated/changed on kms key
This change allows IAM users and roles within the account to update and
replace the tags on the provided KMS key.

If the bootstrap command is run and custom tags are specified, these are
added to the KMS key. If the command is run additional times with
different tags, the key policy needs to allow the replacing / updating
of these tags so that it succeeds.