fix(certificatemanager): unable to set removal policy on DnsValidatedCertificate
This PR adds a method override for `applyRemovalPolicy` which allows the
user to specify a removal policy for the `DnsValidatedCertificate`
construct. Since this construct is backed by a custom resource, the
lambda handler was updated to no longer delete the certificate if the
`RemovalPolicy` is set to `retain`.

This is also needed to allow for an easier migration from
`DnsValidatedCertificate` -> `Certificate`

fixes #20649