fix(integ-tests): can't enable lookups when creating an IntegTest
We were not exposing the `enableLookups` property when creating an
`IntegTest`. This also updates `integ-runner` to ensure that we are
correctly utilizing `enableLooups`. There was an undiscovered (because
you couldn't set `enableLookups=true` :) ) bug which set the dummy
context on _every_ command (i.e. deploy, destroy) when it should have
only been set when synthing the snapshot.