fix(certificatemanager): unable to set removal policy on DnsValidatedCertificate
This PR adds a method override for applyRemovalPolicy which allows the user to specify a removal policy for the DnsValidatedCertificate construct. Since this construct is backed by a custom resource, the lambda handler was updated to no longer delete the certificate if the RemovalPolicy is set to retain.

This is also needed to allow for an easier migration from DnsValidatedCertificate -> Certificate

reroll of #22040
This has the same changes as #22040 with the addition of some logic to
handle only processing updates for certain parameters. If
`RemovalPolicy` is changed for example, the update will not
be processed.

I also added an integration test with some manual instructions. In order
to test ACM certificates I also updated the integ-runner to handle some
additional special env variables.

fixes #20649