fix(s3-deployment): fails when `destinationKeyPrefix` is a token with a long string representation (#22163)
We have a limitation on the length of `destinationKeyPrefix` to be no more than 104 characters. This is because it is used to tag the destination bucket, and tag keys cannot be longer than 128 characters (we prefix the prefix).

However, this validation should not apply when `destinationKeyPrefix` is a token, because in that case it's simply validating the length of the string representation of a token, which doesn't make sense.

----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*