chore: introduce github merit badger workflow (#21898)
See [merit badger readme](https://github.com/kaizencc/github-merit-badger#readme) for an overview of the github action that is introduced in this PR.

The idea is that a "merit badge" will be applied to every incoming PR via label based on prior (successfully merged) contributions to the CDK. The badges are as follows:

- `beginning-contributor`: contributed between 0-2 PRs to the CDK
- `repeat-contributor`: contributed between 3-5 PRs to the CDK
- `valued-contributor`: contributed between 6-12 PRs to the CDK
- `admired-contributor`: contributed between 13-24 PRs to the CDK
- `star-contributor`: contributed between 25-49 PRs to the CDK
- `distinguished-contributor`: contributed 50+ PRs to the CDK

We will **not** label PRs that come from the core team.

Special thanks to @scanlonp and @Tianyi-W if/when this does get merged.