chore(integ-tests): refactor assertions provider
The assertions framework provides utilities for querying data
(e.g. `assertions.awsApiCall`) and separate utilities for making
assertions on that data. Currently these two actions are separated into
separate custom resources so the `awsApiCall` queries data which it
returns and then passes to a _separate_ custom resource which makes
assertions on that data.

Separating these out is unnecessary since we can make the assertion in
the same custom resource. A future PR will introduce the ability to wait
for a certain condition which actually requires making the assertion in
  the same resource.