feat(core): make `StackSynthesizer` easier to subclass
`StackSynthesizer`s used to make heavy use of private APIs, which make
it impossible for people outside the CDK core framework to implement
their own Stack Synthesizers and modify the way assets are referenced.
This question comes up more and more, so this PR opens up the facilities
for doing that.

The changes are as follows:

- Implement `bind()` in `StackSynthesizer`, and add a getter called
  `boundStack` which will return the bound stack.
- Change `synthesizeStackTemplate -> synthesizeTemplate`,
  `emitStackArtifact -> emitArtifact`, both of which will implicitly
  operate on the bound stack.
- Add `addBootstrapVersionRule` so that any subclass can choose to use
  this or not.
- Expose `AssetManifestBuilder`, which stack synthesizers can use to
  build an asset manifest. Refactor the class to do less at the same
  time.
- Expose `cloudFormationLocationFrom...Asset` functions to translate
  a published asset location into a CloudFormation location.
- Document the contract of each method better.

The end result is that it will be easier to mix and match the different
behaviors of the existing synthesizers together into new custom
behavior.

Two new unit test files have been added to show how it is possible
to write ONE intruction to the asset manifest (to upload assets to a
given location) while returning a completely different instruction to
the template (referencing the asset bucket using a CloudFormation
parameter).