fix(core): addPropertyOverride doesn't work for all intrinsics (#22294)
There was a previous fix in #20608 that attempted to fix addPropertyOverride when intrinisics were involved. This PR fixes an edge case where overrides do not work correctly an object is being replaced with an intrinsic.

For example, there might be the case where the override is an intrinsic and it is overriding an object, not a value.

```
   original: {
     Type: 'MyResourceType',
     Properties: {
       prop1: { subprop: { name: { 'Fn::GetAtt': 'abc' } } }
     }
   }
   override: {
     Properties: {
       prop1: { subprop: { 'Fn::If': ['SomeCondition', {...}, {...}] }}
     }
   }
```

The previous fix only handled cases where the original had an intrinsic, but in the above example the override is the first to hit an intrinsic. This PR adds logic to handle cases where we hit an intrinsic in the original _or_ the override.

fixes #19971


----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*