chore(integ-tests): refactor assertions provider (#22238)
The assertions framework provides utilities for querying data (e.g. `assertions.awsApiCall`) and separate utilities for making assertions on that data. Currently these two actions are separated into separate custom resources so the `awsApiCall` queries data which it returns and then passes to a _separate_ custom resource which makes assertions on that data.

Separating these out is unnecessary since we can make the assertion in the same custom resource. A future PR will introduce the ability to wait for a certain condition which actually requires making the assertion in
  the same resource.


----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*