fix(region-info): SSM service principals are incorrect in opt-in regions (#22327)
The change in https://github.com/aws/aws-cdk/pull/17047 was based on an eager reading of the documentation. The documentation referenced only applies to "SSM Inventory", not to "SSM Automation".

And in fact there is no need for that change at all, since all accounts included in the regional service principal are also included in the global service principal.

We therefore revert all use of SSM service principals to the global one.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*