feat(cloudfront-origins): allow setting a user defined origin id
This PR adds the ability to set a custom user defined origin identifier.
The default identifier that is generated is not very descriptive and
some users would like the ability to define a custom identifier.

Since ids have to be unique per distribution, I also added validation
that will throw an error if it is not.

fixes #2756