chore(prlint): cleanup old review comments for less noisy reviews (#22336)
This PR is to quiet down our linter. Currently, it dismisses the review every time an update is made and creating a new review, causing a lot of noise. With this change, it will only create a new review on an initial failure of pr linter and on a new failure after a success if it goes back to failing again. It will create a comment with the failures that will be deleted and rewritten on each run so that only one comment will show up in the PR. Lastly, the review is only dismissed on success.

The lint.ts file can definitely use some refactoring, which I plan to do in a future PR, but this will at least quiet our bot down for the moment. I also plan to stop mergify from dismissing reviews on changes for this context.

----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*