fix(apigateway): cannot create an ApiKey with an imported RestApi
The current implementation of `ApiKey` is a little confusing. It takes a
`resources` property which expects a list of `IRestApi`s. An ApiKey is
not associated with a RestApi by itself though, it needs to be
associated with a specific `Stage` on the RestApi. When you import a
RestApi you cannot import it with a `Stage` and even if you could
provide an `IStage` I don't think that makes the most sense (you can
have multiple stages).

Instead, this PR deprecates the `resources` prop and adds a new `stages`
prop which expects a list of `IStage`s. To facilitate this I have also
done a couple of other things

- Added the ability to import a Stage using
  `Stage.fromStageAttributes()`.
- Added an `addApiKey` method to `IStage`.
- Added additional error handling if the user does try to create an
  ApiKey with an imported RestApi.

fixes #22102