fix(apigateway): CORS OPTIONS method should not require auth
When you create a RestApi and you provide `defaultCorsPreflightOptions`
we automatically create a CORS OPTIONS method for each method. If you
also provide `defaultMethodOptions` then those default options get
passed through to the CORS OPTION method as well. In the case of
authentication options this should not be the case.

This PR explicitly sets the authentication related options to NONE
values which overrides whatever is provided in `defaultMethodOptions`.

I've updated an integration tests to assert that an OPTIONS call is
successful (I also tested before the fix to assert that it failed).

fixes #8615