feat(sns): topic name with `.fifo` suffix generated automatically (#22375)
This PR adds logic for automatic generation of FIFO topics.

When customers don't pass a `topicName` during creation, the construct now generates an unique name that respects the maximum length allowed by the CloudFormation resource (256) and that ends with the `.fifo` suffix.

The previous behavior was to throw an error when the topic was of type FIFO and no name was passed.

Fixes ##18740

----

### All Submissions:

* [x] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [x] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [x] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*