feat(apigateway): support multi-level paths for custom domains
This PR adds support for multi-level paths in api mappings for custom
domains. This is a unique case because in order to create multi-level
mappings for RestApis (ApiGateway v1) you have to use the ApiGateway v2 API.

The aws-apigatewayv2 package is currently an alpha module so this cannot
depend on that module which is why I used the L1 level to implement this
support.

I thought about deprecating the v1 api (BasePathMapping), but that is
still a valid API (and is required if you have an EDGE domain name). The
experience I landed on was to mostly make it transparent to users. When
users create a DomainName, it will now create either a BasePathMapping
or an ApiMapping depending on whether they provide a multi-level
basePath. I did have to introduce a new `addApiMapping` method since
`addBasePathMapping` has a return type of `BasePathMapping`.

I also removed the validation that prevented users from adding
additional basePaths if a (none) basePath was already created. It seems
like that limitation was removed at some point and I have added an
integration test to confirm.

fixes #15904