fix(eks): kubectl layer must contain AWS CLI
The EKS Cluster construct has a property called `kubectlLayer`. When this
property was not given, it would add a two default layers to the custom
resource:

- One with kubectl
- One with the AWS CLI

However, if the property was given, the one layer must contain both
kubectl as well as the AWS CLI.

This makes the `kubectl` layer unnecessarily large -- it must also
contain the AWS CLI which the CDK already has and can bundle itself.

Add a separate `awscliLayer` parameter to control the AWS CLI layer,
if the user so wants. If not, the default AWS CLI layer will be added.

If some user is already using a `kubectlLayer` which includes the AWS
CLI, we now add both the default AWS CLI layer as well as the user's
kubectl layer with the AWS CLI in it. There is no conflict: multiple
layers can contain the same files.  Last layer wins (which is the
kubectl layer with the user's preferred AWS CLI).